changed Long to BigInteger

+++ src/hometasks/Fibonacci/FibNumbers.java
@@ -1,5 +1,6 @@
 package hometasks.Fibonacci;
 
+import java.math.BigInteger;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 
@@ -7,7 +8,7 @@ public class FibNumbers {
 
 	public static void main(String[] args) {
 		System.out.println("Fibonacci numbers:");
-		BlockingQueue<Long> queue = new ArrayBlockingQueue<>(FibNumbersGenerator.ITERATIONS);
+		BlockingQueue<BigInteger> queue = new ArrayBlockingQueue<>(FibNumbersGenerator.ITERATIONS);
 		Thread generator = new Thread(new FibNumbersGenerator(queue));
 		Thread printer = new Thread(new FibNumbersPrinter(queue));
 		generator.start();
